+++ tools/lambda-promtail/lambda-promtail/kinesis.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"context"
+	"time"
+
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/grafana/loki/pkg/logproto"
+	"github.com/prometheus/common/model"
+)
+
+func parseKinesisEvent(ctx context.Context, b *batch, ev *events.KinesisEvent) error {
+	for _, record := range ev.Records {
+		timestamp := time.UnixMilli(record.Kinesis.ApproximateArrivalTimestamp.Unix())
+
+		labels := model.LabelSet{
+			model.LabelName("__aws_log_type"):                 model.LabelValue("kinesis"),
+			model.LabelName("__aws_kinesis_event_source_arn"): model.LabelValue(record.EventSourceArn),
+		}
+
+		labels = applyExtraLabels(labels)
+
+		b.add(ctx, entry{labels, logproto.Entry{
+			Line:      string(record.Kinesis.Data),
+			Timestamp: timestamp,
+		}})
+	}
+
+	return nil
+}
+
+func processKinesisEvent(ctx context.Context, ev *events.KinesisEvent) error {
+	batch, _ := newBatch(ctx)
+
+	err := parseKinesisEvent(ctx, batch, ev)
+	if err != nil {
+		return err
+	}
+
+	err = sendToPromtail(ctx, batch)
+	if err != nil {
+		return err
+	}
+	return nil
+}